refactor(chat): own submission workflow and native commands
